chore: move calculatePostImageAspectRatio to another location
