+++ lib/app/features/feed/views/components/post/components/post_body/components/post_media/post_media.dart
@@ -7,7 +7,7 @@ import 'package:ion/app/features/core/model/media_type.dart';
 import 'package:ion/app/features/feed/views/components/post/components/post_body/components/post_media/components/post_media_carousel.dart';
 import 'package:ion/app/features/feed/views/components/post/components/post_body/components/post_media/components/post_media_item.dart';
 import 'package:ion/app/features/nostr/model/media_attachment.dart';
-import 'package:ion/app/utils/post_media.dart';
+import 'package:ion/app/services/media_service/aspect_ratio.dart';
 
 class PostMedia extends HookConsumerWidget {
   const PostMedia({
